requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
fastapi==0.75.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.1
python-crontab==2.6.0
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
websockets==11.0.3
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0
email-validator==2.0.0